requirements.txt
bs4
chardet
dolreader
qdarkstyle
sortedcontainers
pillow
pygithub
PySide2
pycryptodome
pycryptodome-test-vectors